ETL/Consumers/news_feed_dlx_consumer/requirements.txt
beautifulsoup4==4.12.3
click==8.1.7
joblib==1.4.2
nltk==3.9.1
pika==1.3.2
python-dateutil==2.9.0.post0
regex==2024.9.11
six==1.16.0
soupsieve==2.6
textblob==0.18.0.post0
textblob-fr==0.2.0
tqdm==4.66.5
pymongo==4.10.1
